Alternative algorithm
Instead of always widening the window until we have 3 unambiguous closest
relatives, keep any unambiguous ones you did find, and then widen it for
the rest.

This means that on the site at 24102, you keep Rp22DB159 and RshSTT200,
since they are the best two at the starting window of 50nts. Then you
widen the window for the third.

The original algo widened the window for all of them and ended up
throwing away Rp22DB159 and RshSTT200 and ending up with three BANALs. I
don't see a justification for that other than it's just a bit easier to
code.